+++ packages/windie-sdk-js/cjs/runtime/WindieChatSession.js
@@ -3,12 +3,12 @@
  * Provides the windie chat session module for the TypeScript SDK runtime.
  */
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.WindieChatSession = void 0;
+exports.WindieChatSession = exports.AgentChatSession = void 0;
 const AgentStreamEvents_js_1 = require("./AgentStreamEvents.js");
 function normalizeSendInput(input) {
     return typeof input === 'string' ? { text: input } : input;
 }
-class WindieChatSession {
+class AgentChatSession {
     constructor(conversationRef, runtime) {
         this.conversationRef = conversationRef;
         this.runtime = runtime;
@@ -66,4 +66,5 @@ class WindieChatSession {
         return this.onEvent(listener);
     }
 }
-exports.WindieChatSession = WindieChatSession;
+exports.AgentChatSession = AgentChatSession;
+exports.WindieChatSession = AgentChatSession;